integrate restartice scene to peerconnection scene
