--- Samples~/RestartIce.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9e1f37acfe394490ebd1708f1296b693
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
