--- Samples~/RestartIce/RestartIce.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b6162a25306134a9ea9c9c7556e5f136
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
