--- Samples~/RestartIce/RestartIceSample.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a91bd06c6c95c455fb4493e47a40104f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
